agrego casos de uso para la entidad de cuentas

+++ ms-accounts-neg/src/main/java/com/devsu/challenge/devsu_challenge/contexts/transactions/domain/clazz/Movement.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
@@ -11,8 +13,8 @@ public class Movement {
    private String id;
    private String date;
    private MovementType type;
-   private Double value;
-   private Double balance;
+   private BigDecimal value;
+   private BigDecimal balance;
    private String accountId;
    private AccountType accountType;
 }
